requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dhanhq==2.0.2
fastapi==0.115.8
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
idna==3.10
lxml==5.3.0
motor==3.6.0
multitasking==0.0.11
numpy==2.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.45.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
yfinance==0.2.55
